feat: replace TypeScript MCP server with C#, add 30+ new tools, translate comments to English
- Remove server/ (Node.js/TypeScript) — replaced by revit-mcp-server/ (C# .NET 9)
- Add C# MCP server: Program.cs, RevitToolsHandler, RevitWebSocketClient, DatabaseService
- Add new query tools: element properties, categories, families, worksets, warnings,
  bounding boxes, locations, elements by category/level/view
- Add new modify tools: set parameters, move, rotate, selection, category visibility,
  graphic overrides, element isolation
- Add create_sheet command for sheet creation
- Translate all Chinese comments to English across commandset/ and plugin/
- Rename plugin/Core/Ressources → Resources (fix typo)
- Add new ribbon icons for connection state
- Update root README with full tool listing and C# server instructions
- Update release workflow for C# server build
- Merge upstream transaction mode support for send_code_to_revit

+++ commandset/Services/AnnotationComponents/CreateDimensionEventHandler.cs
@@ -88,7 +88,7 @@ public class CreateDimensionEventHandler : IExternalEventHandler, IWaitableExter
                 View view = null;
                 if (dimInfo.ViewId > 0)
                 {
-                    var element = Doc.GetElement(new ElementId(dimInfo.ViewId));
+                    var element = Doc.GetElement(new ElementId((long)dimInfo.ViewId));
                     view = element as View;
                 }
 
@@ -138,7 +138,7 @@ public class CreateDimensionEventHandler : IExternalEventHandler, IWaitableExter
                             var references = new ReferenceArray();
                             foreach (var elementId in dimInfo.ElementIds)
                             {
-                                var element = Doc.GetElement(new ElementId(elementId));
+                                var element = Doc.GetElement(new ElementId((long)elementId));
                                 if (element != null)
                                 {
                                     // Get appropriate reference for this element
@@ -180,7 +180,7 @@ public class CreateDimensionEventHandler : IExternalEventHandler, IWaitableExter
                             // Apply dimension style if specified
                             if (dimInfo.DimensionStyleId > 0)
                             {
-                                var dimensionType = Doc.GetElement(new ElementId(dimInfo.DimensionStyleId)) as DimensionType;
+                                var dimensionType = Doc.GetElement(new ElementId((long)dimInfo.DimensionStyleId)) as DimensionType;
                                 if (dimensionType != null)
                                 {
                                     dimension.DimensionType = dimensionType;
