+++ commandset/Services/TagRoomsEventHandler.cs
@@ -81,7 +81,7 @@ namespace RevitMCPCommandSet.Services
                 if (_roomIds != null && _roomIds.Count > 0)
                 {
                     // Get level from specified rooms
-                    var firstRoom = _doc.GetElement(new ElementId(_roomIds[0])) as Room;
+                    var firstRoom = _doc.GetElement(new ElementId((long)_roomIds[0])) as Room;
                     if (firstRoom != null)
                     {
                         targetLevel = _doc.GetElement(firstRoom.LevelId) as Level;
@@ -163,7 +163,7 @@ namespace RevitMCPCommandSet.Services
                 {
                     // Get specific rooms by ID
                     rooms = _roomIds
-                        .Select(id => _doc.GetElement(new ElementId(id)))
+                        .Select(id => _doc.GetElement(new ElementId((long)id)))
                         .Where(e => e != null && e is Room)
                         .ToList();
                 }
@@ -374,7 +374,7 @@ namespace RevitMCPCommandSet.Services
             // If specific tag type ID was specified, try to use it
             if (!string.IsNullOrEmpty(_tagTypeId) && int.TryParse(_tagTypeId, out int id))
             {
-                ElementId elementId = new ElementId(id);
+                ElementId elementId = new ElementId((long)id);
                 Element element = doc.GetElement(elementId);
 
                 if (element != null && element is FamilySymbol symbol &&
